Move `get_contract_path` out of abi module
This allows to import it without needing to compile all ABIs, which
in many cases is not necessary.
