+++ raiden/tests/smart_contracts/test_token.py
@@ -3,7 +3,7 @@ import os
 
 from ethereum import tester
 
-from raiden.blockchain.abi import get_contract_path
+from raiden.utils import get_contract_path
 
 
 def test_token():
