+++ tools/init_blockchain.py
@@ -1,7 +1,7 @@
 from pyethapp.rpc_client import JSONRPCClient
 from ethereum.utils import sha3, denoms
 from ethereum._solidity import compile_file
-from raiden.blockchain.abi import get_contract_path
+from raiden.utils import get_contract_path
 from raiden.network.rpc.client import patch_send_transaction
